backend/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pymysql 
python-jose[cryptography] 
passlib[bcrypt] 
pytest
python-multipart
argon2_cffi